Fixes to text rendering system.
Craft needs the bounding box for every 2D entity in order to mark the
dirty regions, so the API of the BetterText was slightly tweaked to allow
the caller to provide the bounding boxes (it seems that I'm unable to
automatically calcualte the bounding boxes, even using the HTML5 APIs for
"measuring text"; not sure if it's a Chrome bug or what).

Also, implemented tooltip for name entry screen.